+++ system.flush.lua
@@ -4,23 +4,30 @@
 -- >> Scripting Wiki for telemeter is wrong. For MaxDistance it is mentioned 20m
 -- >> !!! Telemeter even works (hidden) behind honeycomb
 -- system.print(telemeter.getDistance())
-
-local landing = true
-local surfaceLow = true
 local environmentParameter = getEnvironmentParameter(environmentID)
-landing = telemeter.getDistance() > 0 and telemeter.getDistance() < environmentParameter["surfaceDistanceLanding"]
-surfaceLow = telemeter.getDistance() > 0 and telemeter.getDistance() <= environmentParameter["surfaceDistanceLow"]
 
- 
 
--- Start and landing condition to park on surface
+-- Check if telemeter unit exists
+if telemeter ~= nil then
+    landing = telemeter.getDistance() > 0 and telemeter.getDistance() < environmentParameter["surfaceDistanceLanding"]
+    surfaceLow = telemeter.getDistance() > 0 and telemeter.getDistance() <= environmentParameter["surfaceDistanceLow"]
+end
+
+
+-- Set general altitude stabilization according to the environment
+if surfaceLow and unit.getSurfaceEngineAltitudeStabilization() ~= 0 and
+       unit.getSurfaceEngineAltitudeStabilization() ~= environmentParameter["surfaceDistanceLow"] then
+    unit.activateGroundEngineAltitudeStabilization(environmentParameter["surfaceDistanceLow"])
+end
+    
+
+-- Start and landing condition to stay on surface
 if lockBrake then
     brakeInput = 1
     
     if surfaceLow then
-        Nav.control.setEngineThrust("vertical, thrust", 0) -- default tags are in lower case
-        Nav.axisCommandManager:deactivateGroundEngineAltitudeStabilization()
-        Nav.axisCommandManager:setTargetGroundAltitude(0)
+        unit.setEngineThrust("vertical thrust", 0) -- default tags are in lower case and space seperated
+        unit.deactivateGroundEngineAltitudeStabilization()
     end
 end
 
@@ -39,15 +46,11 @@ end
 
 
 
--- Extend and retract landing gear
-if landing then
-    Nav.control.extendLandingGears()
-else
-    Nav.control.retractLandingGears()
+
+-- In space it is not necessary to use engine stabilization / maneuver engine at high speed
+if environmentID == 2 and velocity_kmh > 2000 then
+    unit.setEngineThrust("vertical thrust", 0)
+    unit.deactivateGroundEngineAltitudeStabilization()
 end
-
-
-
-
 
 -- Darkwinde END: system.flush()